2021 Day 15, Part 2, improved performance
Given that the graph is very sparse this reduces the size of the min_by()
calls, which are innefficient over HashSet. It uses more memory by keeping
an extra HashSet of visited nodes, but it's negligible.

An alternative would be using a Min BinaryHeap, but this solution now runs
in 1 second for Part 2, which is good enough.